requirements.txt
beautifulsoup4==4.13.4
jieba==0.42.1
nltk==3.9.1
openpyxl==3.1.5
pandas==2.2.3
python_docx==1.1.2
rich==14.0.0
xlrd==2.0.1